f32

+++ light/src/light/float.rs
@@ -1,4 +1,4 @@
-pub type Float = f64;
-pub const MIN: f64 = std::f64::MIN;
-pub const MAX: f64 = std::f64::MAX;
-pub const PI: f64 = std::f64::consts::PI;
+pub type Float = f32;
+pub const MIN: f32 = std::f32::MIN;
+pub const MAX: f32 = std::f32::MAX;
+pub const PI: f32 = std::f32::consts::PI;
